poprawki

+++ app/src/main/java/com/triage/model/Victim.java
@@ -18,13 +18,13 @@ public class Victim implements Parcelable, Serializable {
     private boolean changingState;
     private long id;
     private boolean breathing;
-    private int respiratoryRate;
+    private float respiratoryRate;
     private float capillaryRefillTime;
     private boolean walking;
     private TriageColor color;
     private AVPU consciousness;
 
-    public Victim(boolean breathing, int respiratoryRate, float capillaryRefillTime, boolean walking, AVPU consciousness) {
+    public Victim(boolean breathing, float respiratoryRate, float capillaryRefillTime, boolean walking, AVPU consciousness) {
         this.id = totalID; totalID++;
         this.breathing = breathing;
         this.respiratoryRate = respiratoryRate;
@@ -39,7 +39,7 @@ public class Victim implements Parcelable, Serializable {
     }protected Victim(Parcel in) {
         id = in.readLong();
         breathing = (boolean)in.readValue(null);
-        respiratoryRate = in.readInt();
+        respiratoryRate = in.readFloat();
         capillaryRefillTime = in.readFloat();
         walking = (boolean)in.readValue(null);
         color = (TriageColor)in.readValue(null);
@@ -102,11 +102,11 @@ public class Victim implements Parcelable, Serializable {
         this.breathing = breathing;
     }
 
-    public int getRespiratoryRate() {
+    public float getRespiratoryRate() {
         return respiratoryRate;
     }
 
-    public void setRespiratoryRate(int respiratoryRate) {
+    public void setRespiratoryRate(float respiratoryRate) {
         this.respiratoryRate = respiratoryRate;
     }
 
@@ -152,7 +152,7 @@ public class Victim implements Parcelable, Serializable {
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeLong(id);
         dest.writeValue(breathing);
-        dest.writeInt(respiratoryRate);
+        dest.writeFloat(respiratoryRate);
         dest.writeFloat(capillaryRefillTime);
         dest.writeValue(walking);
         dest.writeValue(color);
